+++ test/e2e/framework/BUILD
@@ -14,7 +14,6 @@ go_library(
         "google_compute.go",
         "kubelet_stats.go",
         "log_size_monitoring.go",
-        "metrics_util.go",
         "networking_utils.go",
         "nodes_util.go",
         "perf_util.go",
@@ -48,7 +47,6 @@ go_library(
         "//pkg/master/ports:go_default_library",
         "//pkg/registry/core/service/portallocator:go_default_library",
         "//pkg/scheduler/algorithm/predicates:go_default_library",
-        "//pkg/scheduler/metrics:go_default_library",
         "//pkg/scheduler/nodeinfo:go_default_library",
         "//pkg/security/podsecuritypolicy/seccomp:go_default_library",
         "//pkg/util/system:go_default_library",
@@ -113,7 +111,6 @@ go_library(
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/github.com/onsi/gomega/types:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
-        "//vendor/github.com/prometheus/common/expfmt:go_default_library",
         "//vendor/github.com/prometheus/common/model:go_default_library",
         "//vendor/golang.org/x/net/websocket:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
