+++ test/e2e/framework/flake_reporting_util.go
@@ -22,6 +22,7 @@ import (
 	"sync"
 
 	e2elog "k8s.io/kubernetes/test/e2e/framework/log"
+	e2emetrics "k8s.io/kubernetes/test/e2e/framework/metrics"
 )
 
 // FlakeReport is a struct for managing the flake report.
@@ -90,7 +91,7 @@ func (f *FlakeReport) PrintHumanReadable() string {
 func (f *FlakeReport) PrintJSON() string {
 	f.lock.RLock()
 	defer f.lock.RUnlock()
-	return PrettyPrintJSON(f)
+	return e2emetrics.PrettyPrintJSON(f)
 }
 
 // SummaryKind returns the summary of flake report.
