+++ test/e2e/windows/density.go
@@ -31,6 +31,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2elog "k8s.io/kubernetes/test/e2e/framework/log"
+	e2emetrics "k8s.io/kubernetes/test/e2e/framework/metrics"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 
@@ -49,7 +50,7 @@ var _ = SIGDescribe("[Feature:Windows] Density [Serial] [Slow]", func() {
 				podsNr:   10,
 				interval: 0 * time.Millisecond,
 				// percentile limit of single pod startup latency
-				podStartupLimits: framework.LatencyMetric{
+				podStartupLimits: e2emetrics.LatencyMetric{
 					Perc50: 30 * time.Second,
 					Perc90: 54 * time.Second,
 					Perc99: 59 * time.Second,
@@ -85,12 +86,12 @@ type densityTest struct {
 	// performance limits
 	cpuLimits            framework.ContainersCPUSummary
 	memLimits            framework.ResourceUsagePerContainer
-	podStartupLimits     framework.LatencyMetric
+	podStartupLimits     e2emetrics.LatencyMetric
 	podBatchStartupLimit time.Duration
 }
 
 // runDensityBatchTest runs the density batch pod creation test
-func runDensityBatchTest(f *framework.Framework, testArg densityTest) (time.Duration, []framework.PodLatencyData) {
+func runDensityBatchTest(f *framework.Framework, testArg densityTest) (time.Duration, []e2emetrics.PodLatencyData) {
 	const (
 		podType = "density_test_pod"
 	)
@@ -127,7 +128,7 @@ func runDensityBatchTest(f *framework.Framework, testArg densityTest) (time.Dura
 		firstCreate metav1.Time
 		lastRunning metav1.Time
 		init        = true
-		e2eLags     = make([]framework.PodLatencyData, 0)
+		e2eLags     = make([]e2emetrics.PodLatencyData, 0)
 	)
 
 	for name, create := range createTimes {
@@ -135,7 +136,7 @@ func runDensityBatchTest(f *framework.Framework, testArg densityTest) (time.Dura
 		gomega.Expect(ok).To(gomega.Equal(true))
 
 		e2eLags = append(e2eLags,
-			framework.PodLatencyData{Name: name, Latency: watch.Time.Sub(create.Time)})
+			e2emetrics.PodLatencyData{Name: name, Latency: watch.Time.Sub(create.Time)})
 
 		if !init {
 			if firstCreate.Time.After(create.Time) {
@@ -150,7 +151,7 @@ func runDensityBatchTest(f *framework.Framework, testArg densityTest) (time.Dura
 		}
 	}
 
-	sort.Sort(framework.LatencySlice(e2eLags))
+	sort.Sort(e2emetrics.LatencySlice(e2eLags))
 	batchLag := lastRunning.Time.Sub(firstCreate.Time)
 
 	deletePodsSync(f, pods)
